db/requirements.txt
# ----------------------------
# Data generation & analysis
# ----------------------------
pandas>=2.0
numpy>=2.3.4
Faker>=23.0

# ----------------------------
# Database
# ----------------------------
psycopg2-binary>=2.9
python-dotenv>=1.0